feat: Created the mechanism that shows the sorted list.
Included two for-loops. The first one shows the elements of each box
starting from the beggining to the end of the list. The second for-loop
shows all the elements of the list but starting from the end to the
beggining of it. Thats possible because it is a bidirectional linked
list.